Move all form classes to same package
